+++ src/main/java/repositoriy/UserRepository.java
@@ -1,12 +1,12 @@
 package repositoriy;
 
-import model.User;
+import entity.user.User;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface UserRepository {
     List<User> getAll();
-    Optional<User>getById(long userId);
+    Optional<User>getById(int userId);
     User save(User user);
 }
